requirements.txt
Pillow
pandas
plotly
streamlit
streamlit-folium
folium
tqdm
